Excluding completed orders

+++ scripts/mart_monitor_commenter.py
@@ -222,7 +222,7 @@ def format_results(results: list) -> list:
                     ELSE '🔴'
                 END
                 || ' '
-                || CAST({metric} AS NUMERIC)
+                || CAST({metric} AS INTEGER)
                 || ' ('
                 || ROUND(diff_{metric}_pct, 2)
                 || '%)'
